+++ src/utils/constants.mjs
@@ -1,0 +1,15 @@
+export const mockUsers = [
+    { id: 1, username: "jondo", displayName: "John Doe" },
+    { id: 2, username: "janes", displayName: "Jane Smith" },
+    { id: 3, username: "aljon", displayName: "Alice Johnson" },
+    { id: 4, username: "bobby3", displayName: "Bob Brown" },
+    { id: 5, username: "charl", displayName: "Charlie Davis" },
+];
+
+export const mockProducts = [
+    { id: 1, name: "chicken breast", price: 5.99 },
+    { id: 2, name: "salmon fillet", price: 12.99 },
+    { id: 3, name: "ground beef", price: 8.49 },
+    { id: 4, name: "pork chops", price: 6.99 },
+    { id: 5, name: "tofu", price: 3.49 },
+];
